[ISSUE #50] scan to pay and order complete. (#51)
* 查询缴费支付完成

* C端支付成功,订单查看

* 扫码缴费离场订单记录,优惠券使用完成

* 修改支付库pom引用方式

+++ shared-trad-order-serv/shared-trad-order-impl/src/main/java/cn/suparking/order/service/ChargeInfoService.java
@@ -19,5 +19,5 @@ public interface ChargeInfoService {
      * @param chargeInfoDTO 计费详情信息
      * @return Integer
      */
-    Integer createOrUpdate(ChargeInfoDTO chargeInfoDTO);
+    Long createOrUpdate(ChargeInfoDTO chargeInfoDTO);
 }
